requirements.txt
# Core
numpy>=1.24
pandas>=2.0
pyyaml>=6.0
tqdm>=4.65
pillow>=10.0

# Vision / DL
torch>=2.1
torchvision>=0.16
timm>=1.0.9
ultralytics>=8.3

# Evaluation / viz
scikit-learn>=1.3
matplotlib>=3.7
seaborn>=0.13
opencv-python>=4.8
grad-cam>=1.5.4

# COCO utils (YOLO conversion)
pycocotools>=2.0.7

# Notebooks
jupyter>=1.0
ipykernel>=6.25
